server/requirements.txt
flask
flask-cors
gtts
python-dotenv
requests
git+https://github.com/openai/whisper.git
